Update to set to a random image whenever it runs
We also clean up old files from the image directory so they
don't accumlate infinitely

+++ bingdaily.go
@@ -6,12 +6,15 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"log"
+	"math/rand"
 	"net/http"
 	"os"
 	"os/exec"
 	"os/user"
 	"path"
+	"sort"
 	"strings"
 	"time"
 
@@ -67,14 +70,14 @@ func Execute() error {
 
 	log.Printf("Obtained image metadata: %v\n", im)
 
-	imageName, err := downloadImage(targetDir, im)
+	err = downloadImage(targetDir, im)
 	if err != nil {
 		return fmt.Errorf("failed to download image: %w", err)
 	}
 
-	log.Printf("Setting background image to: %s\n", imageName)
+	log.Println("Updating background image")
 
-	err = setWallpaper(imageName)
+	err = setWallpaper(targetDir)
 	if err != nil {
 		return fmt.Errorf("failed to set wallpaper: %w", err)
 	}
@@ -102,44 +105,74 @@ func getLatestMetadata() (*imageMetadata, error) {
 	return &r.Images[0], nil
 }
 
-func downloadImage(imageDir string, im *imageMetadata) (string, error) {
-	filename := path.Join(imageDir, time.Now().Format("2006-01-02")) + ".jpg"
+func downloadImage(imageDir string, im *imageMetadata) error {
+	filename := path.Join(imageDir, im.Hash+".jpg")
+
+	log.Printf("Checking for existence of file: %s", filename)
+
+	exists, err := imageExists(filename)
+	if err != nil {
+		return fmt.Errorf("unable to determine whether file exists: %v", err)
+	}
+
+	if exists {
+		log.Println("Image already exists, no need to download")
+		return nil
+	}
 
 	fullURL := bingRoot + im.URL
 
 	resp, err := http.Get(fullURL)
 	if err != nil {
-		return "", fmt.Errorf("error while downloading image: %w", err)
+		return fmt.Errorf("error while downloading image: %w", err)
 	}
 
 	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("unexpected response code: %s", resp.Status)
+		return fmt.Errorf("unexpected response code: %s", resp.Status)
 	}
 
 	defer resp.Body.Close()
 
 	file, err := os.Create(filename)
 	if err != nil {
-		return "", fmt.Errorf("failed to create output file: %w", err)
+		return fmt.Errorf("failed to create output file: %w", err)
 	}
 
 	defer file.Close()
 
 	_, err = io.Copy(file, resp.Body)
 	if err != nil {
-		return "", fmt.Errorf("failed to write image to output file: %w", err)
+		return fmt.Errorf("failed to write image to output file: %w", err)
 	}
 
-	return filename, nil
+	return nil
 }
 
-func setWallpaper(filename string) error {
+func imageExists(filename string) (bool, error) {
+	_, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return err == nil, err
+}
+
+func setWallpaper(dirname string) error {
+	err := removeOldFiles(dirname)
+	if err != nil {
+		return fmt.Errorf("Failed to remove old files: %v", err)
+	}
+
+	filename, err := chooseImage(dirname)
+	if err != nil {
+		return fmt.Errorf("Failed to choose an image")
+	}
+
 	dbusAddress, err := obtainDbusAddress()
 	if err != nil {
 		return fmt.Errorf("Failed to obtain dbus address: %w", err)
 	}
 
-	fullFilename := "file://" + filename
+	fullFilename := "file://" + path.Join(dirname, filename)
 
 	log.Printf("Full filename: %s\n", fullFilename)
 
@@ -157,6 +190,51 @@ func setWallpaper(filename string) error {
 		return fmt.Errorf("failed to set wallpaper: %w\n%s", err, out.String())
 	}
 	return nil
+}
+
+type fileSlice []os.FileInfo
+
+func (f fileSlice) Len() int           { return len(f) }
+func (f fileSlice) Swap(i, j int)      { f[i], f[j] = f[j], f[i] }
+func (f fileSlice) Less(i, j int) bool { return f[i].ModTime().Before(f[j].ModTime()) }
+
+func removeOldFiles(dirname string) error {
+	files, err := ioutil.ReadDir(dirname)
+	if err != nil {
+		return fmt.Errorf("Failed to read image directory: %v", err)
+	}
+
+	if len(files) < 10 {
+		log.Println("No images to delete")
+		return nil
+	}
+
+	sort.Sort(fileSlice(files))
+
+	filesToDelete := files[0 : len(files)-10]
+
+	for _, file := range filesToDelete {
+		filename := path.Join(dirname, file.Name())
+		log.Printf("Deleting image: %s\n", filename)
+
+		err = os.Remove(filename)
+		if err != nil {
+			return fmt.Errorf("Failed to delete image: %v", err)
+		}
+	}
+
+	return nil
+}
+
+func chooseImage(dirname string) (string, error) {
+	files, err := ioutil.ReadDir(dirname)
+	if err != nil {
+		return "", fmt.Errorf("Failed to read image directory: %v", err)
+	}
+
+	rand.Seed(time.Now().Unix())
+
+	return files[rand.Intn(len(files))].Name(), nil
 }
 
 func obtainDbusAddress() (string, error) {
